feat: s3 type repository init
